Views and switching views works!!
There is a lot here, but the gist is that a pico specific screen that meets the showdown interface can be created and passed into a number of views on the showdown controller.

Very basic button presses are captured and sent to the controller, which will change which view is being drawn.