Introducing System Info screen.
The system will now determine if it is running on battery power or plugged in charging. The full system is operational either way, and the details of the battery is displayed on a system info screen.

There is still work needed to get the screen to know when to update based on changes to system info.