Added a way for a view to get a callback when the model is updated, skipping the controller.
The good: now the system info is updated any time the system changes state and can directly invoke the correct view. This keeps the controller from needing to know anything special about the views it holds and where they are in the list.

The bad: seems a little complicated... also, the views now need to know if they are "active". Similar to an app or window knowing it's in the foreground or open.